🚧 move Logic in gallery inside parent component (Employees.jsx)
