+++ src/modules/views/Employees/Employees.jsx
@@ -3,16 +3,20 @@ import {Header} from '../../components/Header/Header'
 import {MainContent} from '../../components/MainContent/MainContent'
 import {LateralNav} from '../../components/LateralNav/LateralNav'
 import {EmployeesToolbar} from './EmployeesToolBar/EmployeesToolbar.jsx'
-import {useEffect} from 'react'
+import {useEffect, useState} from 'react'
 import {EmployeesGallery} from './EmployeesGallery/EmployeesGallery'
 import {ViewContext} from '../../../context/EmpoyeesViewContext.jsx'
 import useBoolean from '../../../hooks/useBoolean.jsx'
 import {EmployeesTable} from './EmployeesTable/EmployeesTable'
+import {PaginationLimiter} from '../../components/PaginationLimiter/PaginationLimiter.jsx'
+import {Paginator} from '../../components/Paginator/Paginator.jsx'
+import {usePagination} from '../../../hooks/usePagination.jsx'
+import {useGetAllEmployees} from '../../../api/employees/useGetEmployees.js'
 
 export const Employees = () => {
 	const [tableView, {setToggle: toggleTableView}] = useBoolean(false)
-	const {data} = useQuery(['login'], {enabled: false})
-	const {userInfos} = data
+	const {data: user} = useQuery(['login'], {enabled: false})
+	const {userInfos} = user
 	const {company} = userInfos
 	
 	const setCompanyTheme = () => {
@@ -24,17 +28,39 @@ export const Employees = () => {
 		setCompanyTheme()
 	}, [])
 	
+	const [page, currentPage, firstPage, lastPage, {setPrev, setNext, setPage}] = usePagination()
+	const [limit, setLimit] = useState(12)
+	const {data, isLoading, isError} = useGetAllEmployees(page, limit, {enabled: true})
+	const numberOfPages = Math.ceil(data?.employeesLength / limit)
+	
 	return (
 		<>
 			<Header company={company}/>
 			<MainContent>
 				<LateralNav/>
-				<ViewContext.Provider value={{toggleTableView}}>
-					<section className='main-section'>
-						<EmployeesToolbar/>
-						{tableView ? <EmployeesTable/> : <EmployeesGallery/>}
-					</section>
-				</ViewContext.Provider>
+				{isLoading ? <div>LOADING</div> : isError ? <div>ERROR</div> :
+					<ViewContext.Provider value={{toggleTableView}}>
+						<section className='employees__main-section'>
+							<EmployeesToolbar/>
+							
+							{tableView ? <EmployeesTable/> : <EmployeesGallery employees={data.employees}/>}
+							
+							<div className='employees__pagination-container'>
+								<PaginationLimiter setLimit={setLimit} text='employees' totalData={data.employeesLength} currentPage={currentPage}/>
+								<p className='employees__totalFound'><span>{data.employeesLength}</span>{data.employeesLength > 1 ? ' Employees' : ' Employee'}</p>
+								<Paginator totalOfPages={numberOfPages}
+								           setPage={setPage}
+								           currentPage={currentPage}
+								           firstPage={firstPage}
+								           lastPage={lastPage(numberOfPages)}
+								           setPrev={setPrev}
+								           setNext={setNext}
+								/>
+							</div>
+						
+						</section>
+					</ViewContext.Provider>
+				}
 			</MainContent>
 		</>
 	)
